+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# next
+# 3.8.6
 - Fix bug collision. [#511](https://github.com/RafaelBarbosatec/bonfire/issues/511)
 - Renamed `AutomaticRandomMovement` to `RandomMovement`
 - `RandomMovement` improvements! Now works in `PLatformEnemy`